Initial commit enabling compilation with gmake on CURRENT
Preliminary/hack patch that will get things building. Not tested that
the produced binaries actually work.

To fix the str + int warning, used the same approach as in Qemu:
http://lists.nongnu.org/archive/html/qemu-devel/2013-08/msg00537.html

Removed a few warnings that were not supported by our clang
(-Wmissing-parameter-type, -Wold-style-declaration. Disabled
warnings where required for uninitialized vars and unused functions,
and a -Wno-dangling-else as a quick fix for a few if statements that
don't have brackets.

Signed-off-by: Vassilis Laganakos <vassilis@einval.com>